move tx register host zone
