+++ x/stakeibc/client/cli/tx.go
@@ -2,6 +2,7 @@ package cli
 
 import (
 	"fmt"
+	"strconv"
 	"time"
 
 	errorsmod "cosmossdk.io/errors"
@@ -14,6 +15,13 @@ import (
 	"github.com/spf13/cobra"
 
 	"github.com/Stride-Labs/stride/v22/x/stakeibc/types"
+)
+
+const (
+	FlagMinRedemptionRate            = "min-redemption-rate"
+	FlagMaxRedemptionRate            = "max-redemption-rate"
+	FlagCommunityPoolTreasuryAddress = "community-pool-treasury-address"
+	FlagMaxMessagesPerIcaTx          = "max-messages-per-ica-tx"
 )
 
 var DefaultRelativePacketTimeoutTimestamp = cast.ToUint64((time.Duration(10) * time.Minute).Nanoseconds())
@@ -113,6 +121,97 @@ func CmdLSMLiquidStake() *cobra.Command {
 	}
 
 	flags.AddTxFlagsToCmd(cmd)
+
+	return cmd
+}
+
+func CmdRegisterHostZone() *cobra.Command {
+	cmd := &cobra.Command{
+		Use:   "register-host-zone [connection-id] [host-denom] [bech32prefix] [ibc-denom] [channel-id] [unbonding-period] [lsm-enabled]",
+		Short: "Broadcast message register-host-zone",
+		Args:  cobra.ExactArgs(7),
+		RunE: func(cmd *cobra.Command, args []string) error {
+			clientCtx, err := client.GetClientTxContext(cmd)
+			if err != nil {
+				return err
+			}
+
+			connectionId := args[0]
+			hostDenom := args[1]
+			bech32prefix := args[2]
+			ibcDenom := args[3]
+			channelId := args[4]
+			unbondingPeriod, err := strconv.ParseUint(args[5], 10, 64)
+			if err != nil {
+				return err
+			}
+			lsmEnabled, err := strconv.ParseBool(args[6])
+			if err != nil {
+				return err
+			}
+
+			minRedemptionRateStr, err := cmd.Flags().GetString(FlagMinRedemptionRate)
+			if err != nil {
+				return err
+			}
+			minRedemptionRate := sdk.ZeroDec()
+			if minRedemptionRateStr != "" {
+				minRedemptionRate, err = sdk.NewDecFromStr(minRedemptionRateStr)
+				if err != nil {
+					return err
+				}
+			}
+
+			maxRedemptionRateStr, err := cmd.Flags().GetString(FlagMaxRedemptionRate)
+			if err != nil {
+				return err
+			}
+			maxRedemptionRate := sdk.ZeroDec()
+			if maxRedemptionRateStr != "" {
+				maxRedemptionRate, err = sdk.NewDecFromStr(maxRedemptionRateStr)
+				if err != nil {
+					return err
+				}
+			}
+
+			communityPoolTreasuryAddress, err := cmd.Flags().GetString(FlagCommunityPoolTreasuryAddress)
+			if err != nil {
+				return err
+			}
+
+			maxMessagesPerIcaTx, err := cmd.Flags().GetUint64(FlagMaxMessagesPerIcaTx)
+			if err != nil {
+				return err
+			}
+
+			msg := types.NewMsgRegisterHostZone(
+				clientCtx.GetFromAddress().String(),
+				connectionId,
+				bech32prefix,
+				hostDenom,
+				ibcDenom,
+				channelId,
+				unbondingPeriod,
+				minRedemptionRate,
+				maxRedemptionRate,
+				lsmEnabled,
+				communityPoolTreasuryAddress,
+				maxMessagesPerIcaTx,
+			)
+
+			if err := msg.ValidateBasic(); err != nil {
+				return err
+			}
+
+			return tx.GenerateOrBroadcastTxCLI(clientCtx, cmd.Flags(), msg)
+		},
+	}
+
+	flags.AddTxFlagsToCmd(cmd)
+	cmd.Flags().String(FlagMinRedemptionRate, "", "minimum redemption rate")
+	cmd.Flags().String(FlagMaxRedemptionRate, "", "maximum redemption rate")
+	cmd.Flags().String(FlagCommunityPoolTreasuryAddress, "", "community pool treasury address")
+	cmd.Flags().Uint64(FlagMaxMessagesPerIcaTx, 0, "maximum number of ICA txs in a given tx")
 
 	return cmd
 }
